% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compliment}
\alias{compliment}
\title{Random compliment}
\usage{
compliment(target)
}
\arguments{
\item{target}{the target of the compliment}
}
\description{
generates and returns a random comment from
http://peoplearenice.blogspot.com/p/compliment-list.html
}

