% Generated by roxygen2 (4.0.1): do not edit by hand
\name{frontDoor}
\alias{frontDoor}
\title{Main Looping Function, called for every statement in a channel}
\usage{
frontDoor(env)
}
\arguments{
\item{env}{the Rook environment}
}
\description{
This is the main input function for anything in the channel.
This is a rook app function.
}

