% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inspirational}
\alias{inspirational}
\title{Random inspirational quote}
\usage{
inspirational()
}
\description{
generates and returns a random inspirational quote from
http://www.forbes.com/sites/kevinkruse/2013/05/28/inspirational-quotes/
}

