% Generated by roxygen2 (4.0.1): do not edit by hand
\name{insult}
\alias{insult}
\title{Returns a random insult}
\usage{
insult(user, from)
}
\arguments{
\item{user}{the name of the entity to insult}

\item{from}{the person asking for the insult}
}
\description{
Given the user to insult (and occasionally the person wanting to insult)
This will generate a random insult and return the string.
}

